LayoutTests/fast/history/page-cache-clearing-expected.txt

Tests that clearing the page cache doesn't leave pages in the cache.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


pageshow - not from cache
pagehide - entering cache
pageshow - from cache
Cached page count before clearing: 1
Clearing the page cache
Cached page count after clearing: 0
PASS
